XGBoost-API/requirements.txt
streamlit
pandas
joblib
xgboost
azure-cosmos==4.0.0
scikit-learn
plotly
Image
